RM-2217 Upgrade Aikau and use new encode URI component for parameters.
The behaviour of the search filter on the "Security Clearance" page is
now very similar to the "Users" page.  The two notable exceptions are
searching for a space (' ') and searching for an asterisk ('*'). Both
of these produce errors on the "Users" page, and return all users on
the "Security Clearance" page.

git-svn-id: https://svn.alfresco.com/repos/alfresco-enterprise/modules/recordsmanagement/HEAD@108176 c4b6b30b-aa2e-2d43-bbcb-ca4b014f7261