RM-2227 Remove commons-logging dependency from Share amp.
This should mean we no longer need to install the AMP with "-force".

Locally the issue was fixed by changing the ${mvn.exec} definition in
build.local.properties to say:
mvn.exec=mvn
And ensuring my system version of maven is a recent version (3.2.5 rather
than 3.0.5 which I was using by accident before).

+review RM

git-svn-id: https://svn.alfresco.com/repos/alfresco-enterprise/modules/recordsmanagement/HEAD@111551 c4b6b30b-aa2e-2d43-bbcb-ca4b014f7261