ACE-4558: API Share and Platform builds should conatin the API change tests
The AlfrescoPublicApi annotation has been removed from the classes in the Share codebase as the annotation is supposed to represent classes in the platform. These classes will still remain part of the public API but as an official Share Extension Point, as such they will be tested via an extension point sample.