Make sure artifacts look up is done first on Maven Central
While looking up for artifacts to download, Maven will use first the repositories defined in the settings and if it's not found it will default to Maven Central. Considering that most of the dependencies are coming from Central it's better to put Maven Central in the first place in the repository list to reduce the amount of time spent looking for the artifact in the wrong repository.