Unique name for packaging/docker-alfresco/pom.xml (#100)
The packaging/docker-alfresco module was copied from acs-packaging, but the name was not changed. As a result we now have issues in acs-packaging as it is picking up the wrong version (7.0.0) of the rest-api-explorer in the tomcat dev env. It should be using the latest one defined in alfresco-communit-repo/pom.xml.