[ACS-9045] Bump Spring to 6.2.0 (#860)
Bumping Spring Framework to 6.2.0 to enable Spring upgrade in ACS.

Due to removal of useExpiresHeader and useCacheControlHeader from WebContentGenerator.java it was impossible to upgrade Spring to 6.2.X in ACS since surf-webscripts was still requiring those properties.

As per WebContentGenerator.java documentation:

/**
* Set whether to use the HTTP 1.0 expires header. Default is "false",
* as of 4.2.
* <p>Note: Cache headers will only get applied if caching is enabled
* (or explicitly prevented) for the current request.
* @deprecated as of 4.2, since going forward, the HTTP 1.1 cache-control
* header will be required, with the HTTP 1.0 headers disappearing
*/
Those two are deprecated and were removed in 6.2.0. They are no longer needed as HTTP 1.0 is not longer supported by Spring.